[orc-rt] Hoist DirectCaller test utility into header to enable re-use.
The DirectCaller utility allows "direct" calls (with arguments serialized into,
and then immediately back out of a WrapperFunctionBuffer) to wrapper functions.
It was introduced for the SPSWrapperFunction tests, but will be useful for
testing WrapperFunction interfaces for various orc-rt APIs too, so this commit
hoists it somewhere where it can be reused.
